pyproject.toml
[project]
name = "gigabot"
version = "0.1.0"
description = "AI-агент на базе GigaChat: управление проектами, документами, RAG, голос, OCR"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GigaBot contributors"}
]
keywords = ["ai", "agent", "gigachat", "telegram", "rag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "typer>=0.20.0,<1.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "loguru>=0.7.3,<1.0.0",
    "rich>=14.0.0,<15.0.0",
    "prompt-toolkit>=3.0.50,<4.0.0",
    "json-repair>=0.57.0,<1.0.0",

    # GigaChat SDK
    "gigachat>=0.2.0,<0.3.0",

    # Telegram
    "python-telegram-bot[socks]>=22.0,<23.0",
    "socksio>=1.0.0,<2.0.0",

    # RAG
    "chromadb>=0.5.0,<1.0.0",

    # Documents
    "python-docx>=1.0.0,<2.0.0",
    "openpyxl>=3.1.0,<4.0.0",
    "pdfplumber>=0.11.0,<0.12.0",

    # OCR
    "pytesseract>=0.3.10,<1.0.0",
    "Pillow>=10.0.0,<12.0.0",

    # SaluteSpeech
    "httpx>=0.28.0,<1.0.0",

    # Web
    "readability-lxml>=0.8.4,<1.0.0",

    # Scheduling
    "croniter>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
gigabot = "gigabot.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gigabot"]

[tool.hatch.build.targets.wheel.sources]
"gigabot" = "gigabot"

[tool.hatch.build]
include = [
    "gigabot/**/*.py",
    "gigabot/skills/**/*.md",
]

[tool.hatch.build.targets.sdist]
include = [
    "gigabot/",
    "README.md",
    "LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
